// @form/form/models/form.model.ts
export interface JsonFormValidators {
    min?: number;
    max?: number;
    required?: boolean;
    requiredTrue?: boolean;
    email?: boolean;
    minLength?: boolean;
    maxLength?: boolean;
    pattern?: string;
    nullValidator?: boolean;
    
  }
  export interface JsonFormControlOptions {
    min?: string;
    max?: string;
    step?: string;
    icon?: string;
  }
  export  interface JsonFormControls {
    name: string;
    label: string;
    value: string;
    type: string;
    selectOptions?:string[];
    hasAsyncSearch?:boolean;
    asyncSelectOptions?:string;
    checkBoxGroup?:any;
    radioGroup?:any;
    options?: JsonFormControlOptions;
    required: boolean;
    validators: JsonFormValidators;
    hasAsyncValidators?:boolean;
    class?:string;
    placeholder?:string | null;
  }
  export interface JsonFormData {
    controls: JsonFormControls[];
  }